requirements.txt
requests==2.23.0
pyproj==2.6.0
pytest-remotedata==0.3.2
h5json==1.1.3
h5py==2.10.0
munch==2.5.0
numpy==1.18.4
